Reorganize crate
